Changes to price

+++ src/Data/DevCamp.Data.Models/Listing.cs
@@ -17,7 +17,7 @@
 
         public string ProjectDetails { get; set; }
 
-        public double? StartingPrice { get; set; }
+        public decimal? StartingPrice { get; set; }
 
         public bool IsComplete { get; set; }
